third_party/WebKit/LayoutTests/inspector/sources/debugger/source-frame-inline-breakpoint-decorations.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>
function foo()
{
    var p = Promise.resolve().then(() => console.log(42))
        .then(() => console.log(239));
    return p;
}
//# sourceURL=foo.js
</script>
<script>

function test() {
  function waitAndDumpDecorations(sourceFrame) {
    return SourcesTestRunner.waitJavaScriptSourceFrameBreakpoints(sourceFrame)
        .then(() => SourcesTestRunner.dumpJavaScriptSourceFrameBreakpoints(sourceFrame));
  }

  Bindings.breakpointManager._storage._breakpoints = {};
  SourcesTestRunner.runDebuggerTestSuite([
    function testAddRemoveBreakpoint(next) {
      var javaScriptSourceFrame;
      SourcesTestRunner.showScriptSource('foo.js', addBreakpoint);

      function addBreakpoint(sourceFrame) {
        javaScriptSourceFrame = sourceFrame;
        TestRunner.addResult('Setting breakpoint');
        SourcesTestRunner.createNewBreakpoint(javaScriptSourceFrame, 3, '', true)
            .then(() => waitAndDumpDecorations(javaScriptSourceFrame).then(removeBreakpoint));
      }

      function removeBreakpoint() {
        TestRunner.addResult('Toggle breakpoint');
        waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
        SourcesTestRunner.toggleBreakpoint(javaScriptSourceFrame, 3);
      }
    },

    function testAddRemoveBreakpointInLineWithOneLocation(next) {
      var javaScriptSourceFrame;
      SourcesTestRunner.showScriptSource('foo.js', addBreakpoint);

      function addBreakpoint(sourceFrame) {
        javaScriptSourceFrame = sourceFrame;
        TestRunner.addResult('Setting breakpoint');
        SourcesTestRunner.createNewBreakpoint(javaScriptSourceFrame, 5, '', true)
            .then(() => waitAndDumpDecorations(javaScriptSourceFrame).then(removeBreakpoint));
      }

      function removeBreakpoint() {
        TestRunner.addResult('Toggle breakpoint');
        SourcesTestRunner.toggleBreakpoint(javaScriptSourceFrame, 5);
        waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
      }
    },

    function clickByInlineBreakpoint(next) {
      var javaScriptSourceFrame;
      SourcesTestRunner.showScriptSource('foo.js', addBreakpoint);

      function addBreakpoint(sourceFrame) {
        javaScriptSourceFrame = sourceFrame;
        TestRunner.addResult('Setting breakpoint');
        SourcesTestRunner.createNewBreakpoint(javaScriptSourceFrame, 3, '', true)
            .then(() => waitAndDumpDecorations(javaScriptSourceFrame).then(clickBySecondLocation));
      }

      function clickBySecondLocation() {
        TestRunner.addResult('Click by second breakpoint');
        waitAndDumpDecorations(javaScriptSourceFrame).then(clickByFirstLocation);
        SourcesTestRunner.clickJavaScriptSourceFrameBreakpoint(javaScriptSourceFrame, 3, 1, next);
      }

      function clickByFirstLocation() {
        TestRunner.addResult('Click by first breakpoint');
        waitAndDumpDecorations(javaScriptSourceFrame).then(clickBySecondLocationAgain);
        SourcesTestRunner.clickJavaScriptSourceFrameBreakpoint(javaScriptSourceFrame, 3, 0, next);
      }

      function clickBySecondLocationAgain() {
        TestRunner.addResult('Click by second breakpoint');
        waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
        SourcesTestRunner.clickJavaScriptSourceFrameBreakpoint(javaScriptSourceFrame, 3, 1, next);
      }
    },

    function toggleBreakpointInAnotherLineWontRemoveExisting(next) {
      var javaScriptSourceFrame;
      SourcesTestRunner.showScriptSource('foo.js', addBreakpoint);

      function addBreakpoint(sourceFrame) {
        javaScriptSourceFrame = sourceFrame;
        TestRunner.addResult('Setting breakpoint in line 4');
        SourcesTestRunner.toggleBreakpoint(sourceFrame, 4, false);
        waitAndDumpDecorations(javaScriptSourceFrame).then(toggleBreakpointInAnotherLine);
      }

      function toggleBreakpointInAnotherLine() {
        TestRunner.addResult('Setting breakpoint in line 3');
        waitAndDumpDecorations(javaScriptSourceFrame).then(removeBreakpoints);
        SourcesTestRunner.toggleBreakpoint(javaScriptSourceFrame, 3, false);
      }

      function removeBreakpoints() {
        TestRunner.addResult('Click by first inline breakpoints');
        waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
        SourcesTestRunner.clickJavaScriptSourceFrameBreakpoint(javaScriptSourceFrame, 3, 0, next);
        SourcesTestRunner.clickJavaScriptSourceFrameBreakpoint(javaScriptSourceFrame, 4, 0, next);
      }
    }
  ]);
};
</script>
</head>
<body onload="runTest()">
<p>Checks that JavaScriptSourceFrame show inline breakpoints correctly</p>
</body>
</html>
